Update dependencies
This is a very minor patch to raise the minimum constraint on "get-uri".
Previous to 2.0.3, the version restrictions in it's package.json meant
it could have pulled in a vulnerable version of "extend" in certain
circumstances. Raising it ensures NPM can't dedupe a vulnerability into
the transitive dependencies.

See https://github.com/TooTallNate/node-get-uri/pull/12 for more detail.